requirements.txt
# EpiZone — Dependances Python (version PC)
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
pyyaml>=6.0
geopandas>=0.14
shapely>=2.0
pyarrow>=14.0
pyproj>=3.6
dash>=2.14
dash-leaflet>=1.0
dash-bootstrap-components>=1.5
dash-extensions>=1.0
